made a copy of elevator subsystem, then I cut down the original subsystem class as much as possible.
