+++ src/main/java/frc/robot/subsystems/ElevatorSubsystem.java
@@ -12,7 +12,6 @@ import frc.robot.Constants.ElevatorConstants;
 
 public class ElevatorSubsystem extends SubsystemBase {
     private final Encoder heightEncoder = NeckConstants.primaryNeckEncoder;
-    private final int encoderCountsPerRot = 2048;
 
     private final PWMSparkMax elevatorMotor = ClimberConstants.climberMotor;
 
@@ -23,83 +22,12 @@ public class ElevatorSubsystem extends SubsystemBase {
                                                                 ElevatorConstants.ELEVATOR_PID.kI,
                                                                 ElevatorConstants.ELEVATOR_PID.kD);
 
-    private static final double[] setpoints = {0, 0, 3, 11.5};
-
-    private boolean pidOn = true;
-
-    private double getElevatorLevelSetpoint(ElevatorConstants.ElevatorLevel level) {
-        switch (level){
-            case L1:
-            return setpoints[0];
-            case L2:
-            return setpoints[1];
-            case L3:
-            return setpoints[2];
-            case L4:
-            return setpoints[3];
-            default:
-            return -1;
-        }
-    }
-
-    public void setPIDEnabled(boolean enabled) {
-        pidOn = enabled;
-    }
-
-    public void raiseElevatorToLevel(ElevatorConstants.ElevatorLevel level) {
-        elevatorPID.setSetpoint(getElevatorLevelSetpoint(level));
-    }
-
-    /**
-     * Gets the normalized encoder value.
-     * @return The encoder reading, in rotations, with increasing positive meaning increasing elevator height.
-     */
-    public double getEncoderValue() {
-        return heightEncoder.get() / encoderCountsPerRot * -1;
-    }
-    
-    /**
-     * Resets the encoder count to 0.
-     */
-    public void resetEncoder() {
-        System.out.println("Resetting Encoder");
-        heightEncoder.reset();
-    }
-
-    /**
-     * Sets the PID height setpoint of the elevator.
-     * @param setpoint The setpoint height, in normalized encoder revolutions, that the elevator will attempt to reach.
-     */
-    public void setSetpoint(double setpoint) {
-        elevatorPID.setSetpoint(setpoint);
-    }
-
-    public double getSetpoint() {
-        return elevatorPID.getSetpoint();
-    }
-
-    private PIDController getCurrentPIDController() {
-        return elevatorPID;
-    }
-
     public void moveElevatorUp() {
-        if (pidOn) {
-            elevatorPID.setSetpoint(elevatorPID.getSetpoint() + ElevatorConstants.elevatorManualMovementSpeed);
-        }
-        else {
-            elevatorMotor.set(-1);
-            setSetpoint(getEncoderValue());
-        }
+        elevatorPID.setSetpoint(elevatorPID.getSetpoint() + ElevatorConstants.elevatorManualMovementSpeed);
     }
 
     public void moveElevatorDown() {
-        if (pidOn) {
-            elevatorPID.setSetpoint(elevatorPID.getSetpoint() - ElevatorConstants.elevatorManualMovementSpeed);
-        }
-        else {
-            elevatorMotor.set(1);
-            setSetpoint(getEncoderValue());
-        }
+        elevatorPID.setSetpoint(elevatorPID.getSetpoint() - ElevatorConstants.elevatorManualMovementSpeed);
     }
 
     public boolean isTopLimitSwitchPressed() {
@@ -110,33 +38,11 @@ public class ElevatorSubsystem extends SubsystemBase {
         return !bottomLimitSwitch.get();
     }
 
-    public Command configureSetpointForTopOfElevatorCommand() {
-        return runOnce(() -> setSetpoint(heightEncoder.get()));
-    }
-
-    public Command togglePIDEnabledCommand() {
-        return runOnce(() -> setPIDEnabled(!pidOn));
-    }
-
-    public Command moveElevatorUpCommand() {
-        return runOnce(() -> moveElevatorUp());
-    }
-
-    public Command moveElevatorDownCommand() {
-        return runOnce(() -> moveElevatorDown());
-    }
-
-    public Command resetEncoderCommand() {
-        return runOnce(() -> resetEncoder());
-    }
-
-    public void maintainLevel() {
-        elevatorMotor.set(0);
-        if (!pidOn) return;
-        //elevatorMotor.set(elevatorPID.calculate(heightEncoder.get()));
-    }
-
-    /**PAY ATTENTION!!!!! This method assumes that moving the elevator up is done by setting the motor speed to a positive value, and assumes that a higher encoder value means that the elevator is higher.*/
+    /**ATTENTION!<p>
+     * This method (and those it calls) assumes that moving the elevator up is done by
+     * setting the motor speed to a negative value,
+     * and assumes that a lowers encoder value means that the elevator is higher.
+     */
     public void runPID() {
 
         if (isTopLimitSwitchPressed() && isBottomLimitSwitchPressed()) {
@@ -145,53 +51,60 @@ public class ElevatorSubsystem extends SubsystemBase {
         }
 
         if (isTopLimitSwitchPressed()) {
-
-            if (getSetpoint() > getEncoderValue()) {
-                setSetpoint(getEncoderValue());
-            }
-            // NOTE!!!! This if statement assumes that moving the elevator up is done by setting the motor speed to a negative value
-            if (getCurrentPIDController().calculate(getSetpoint()) < 0) {
-                elevatorMotor.set(0);
-            }
-            else {
-                elevatorMotor.set(-getCurrentPIDController().calculate(getEncoderValue()));
-            }
-            return;
+            handleTopLSPressed();
         }
-
-        if (isBottomLimitSwitchPressed()) {
-            resetEncoder();
-            if (getSetpoint() < getEncoderValue()) {
-                setSetpoint(getEncoderValue());
-            }
-            // NOTE!!!! This if statement assumes that moving the elevator down is done by setting the motor speed to a positive value
-            if (getCurrentPIDController().calculate(getSetpoint()) > 0) {
-                elevatorMotor.set(0);
-            }
-            else {
-                elevatorMotor.set(-getCurrentPIDController().calculate(getEncoderValue()));
-            }
-            return;
+        else if (isBottomLimitSwitchPressed()) {
+            handleBottomLSPressed();
         }
-        System.out.println("Here!");
-        elevatorMotor.set(-getCurrentPIDController().calculate(getEncoderValue()));
+        else {
+            elevatorMotor.set(elevatorPID.calculate(heightEncoder.get())); 
+        }
+    }
+
+    private void handleTopLSPressed() {
+        if (elevatorPID.getSetpoint() < heightEncoder.get()) {
+            elevatorPID.setSetpoint(heightEncoder.get());
+        }
+        // NOTE!!!! This if statement assumes that moving the elevator up is done by setting the motor speed to a negative value
+        if (elevatorPID.calculate(heightEncoder.get()) < 0) {
+            elevatorMotor.set(0);
+        }
+        else {
+            elevatorMotor.set(elevatorPID.calculate(heightEncoder.get()));
+        }
+    }
+
+    private void handleBottomLSPressed() {
+        heightEncoder.reset();
+
+        if (elevatorPID.getSetpoint() > heightEncoder.get()) {
+            elevatorPID.setSetpoint(heightEncoder.get());
+        }
+        // NOTE!!!! This if statement assumes that moving the elevator down is done by setting the motor speed to a positive value
+        if (elevatorPID.calculate(heightEncoder.get()) > 0) {
+            elevatorMotor.set(0);
+        }
+        else {
+            elevatorMotor.set(elevatorPID.calculate(heightEncoder.get()));
+        }
     }
 
     // This method is being used to run safety code which should be executed, no matter what.
     public void periodic() {
-        System.out.println("PID On? "+pidOn+", Motor Speed: "+elevatorMotor.get()+", Setpoint:"+getSetpoint()+", encoder: "+getEncoderValue()+", PID output: "+elevatorPID.calculate(getEncoderValue()));
-        System.out.println(isTopLimitSwitchPressed());
-        
-        // NOTE!!!! This if statement assumes that moving the elevator up is done by setting the motor speed to a negative value
-        if (isTopLimitSwitchPressed() && elevatorMotor.get() < 0) {
-            System.out.println("Here");
-            elevatorMotor.set(0);
-        }
-           
-        // NOTE!!!! This if statement assumes that moving the elevator down is done by setting the motor speed to a positive value
-        if (isBottomLimitSwitchPressed() && elevatorMotor.get() > 0) {
-            elevatorMotor.set(0);
-        }
-        
+        System.out.println( "Motor Speed: "+elevatorMotor.get()+
+                            ", Setpoint:"+elevatorPID.getSetpoint()+
+                            ", encoder: "+heightEncoder.get()+
+                            ", PID output: "+elevatorPID.calculate(heightEncoder.get())
+                            );
+    }
+     
+    // Factory Command Section: Add all command factory methods beneath this line.
+
+    public Command moveElevatorUpCommand() {
+        return runOnce(() -> moveElevatorUp());
+    }
+
+    public Command moveElevatorDownCommand() {
+        return runOnce(() -> moveElevatorDown());
     }
 }
